Add option --save-real-time-to
Option may be used to save real time of the simulation to the file
specified by the user. It is necessary when the input consists of
many files (subfiles) [option --root-many] and some of them may be
broken [for example not closed properly].
Then, in case of each file, time of the simulation is calculated as
a difference between the last and the first hit in the subfile.
After calculating coincidences for all subfiles, times are summed
and the final time is saved.